+++ admin/delete_product.php
@@ -3,10 +3,17 @@
   include_once("../config/db.php");
   
   if(!$con = mysqli_connect($dbhost, $dbuser, $dbpass, $dbname)) {
-    die("failed to connect!");
+    echo "Failed to connect!";
+    die();
   }
 
   $id = $_POST['id'];
+
+  if (!isset($id)) {
+    echo "Invalid input";
+    die;
+  }
+
   $query = "delete from product where id = $id";
 
   $result = mysqli_query($con, $query);
